pyproject.toml
[project]
name = "memex"
version = "0.1.0"
description = "A second brain you talk to like your first"
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"
authors = [{ name = "Brendan Cooley" }]
dependencies = [
    "pydantic-ai>=0.1.0",
    "anthropic>=0.25.0",
    "diskcache>=5.6",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/memex"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/memex"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[dependency-groups]
dev = [
    "mkdocs-material>=9.7.1",
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.9",
    "ty>=0.0.12",
]
